pc/requirements-training.txt
# Recommended: use a dedicated Python 3.10-3.12 virtual environment on Windows for training.
torch>=2.6
torchvision>=0.21
numpy>=1.24
Pillow>=10.0
PyYAML>=6.0
tqdm>=4.66
onnx>=1.18
onnxscript>=0.5
onnxruntime>=1.20
cmake>=4.0
ninja>=1.11
